ai-services/requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
langchain==0.1.6
langchain-openai==0.0.5
langchain-community==0.0.18
langgraph==0.0.20
sqlalchemy==2.0.25
psycopg2-binary==2.9.9
pgvector==0.2.4
openai==1.10.0
tiktoken==0.5.2
python-dotenv==1.0.0
httpx==0.26.0
pydantic==2.5.3
pydantic-settings==2.1.0

# Document processing dependencies
PyPDF2==3.0.1
python-docx==1.1.0
beautifulsoup4==4.12.3
lxml==5.1.0

# File upload support
python-multipart==0.0.6
